refactor(web): remove redundant dataset card-item components and related code
