+++ web/app/components/app/configuration/dataset-config/card-item/index.tsx
@@ -1,58 +1,112 @@
 'use client'
 import type { FC } from 'react'
-import React from 'react'
+import React, { useState } from 'react'
+import {
+  RiDeleteBinLine,
+  RiEditLine,
+} from '@remixicon/react'
 import { useTranslation } from 'react-i18next'
-import TypeIcon from '../type-icon'
-import RemoveIcon from '../../base/icons/remove-icon'
-import s from './style.module.css'
-import cn from '@/utils/classnames'
+import SettingsModal from '../settings-modal'
 import type { DataSet } from '@/models/datasets'
-import { formatNumber } from '@/utils/format'
-import Tooltip from '@/app/components/base/tooltip'
+import ActionButton, { ActionButtonState } from '@/app/components/base/action-button'
+import Drawer from '@/app/components/base/drawer'
+import useBreakpoints, { MediaType } from '@/hooks/use-breakpoints'
+import Badge from '@/app/components/base/badge'
+import { useKnowledge } from '@/hooks/use-knowledge'
+import cn from '@/utils/classnames'
+import AppIcon from '@/app/components/base/app-icon'
 
-export type ICardItemProps = {
+type ItemProps = {
   className?: string
   config: DataSet
   onRemove: (id: string) => void
   readonly?: boolean
+  onSave: (newDataset: DataSet) => void
+  editable?: boolean
 }
-const CardItem: FC<ICardItemProps> = ({
-  className,
+
+const Item: FC<ItemProps> = ({
   config,
+  onSave,
   onRemove,
-  readonly,
+  editable = true,
 }) => {
+  const media = useBreakpoints()
+  const isMobile = media === MediaType.mobile
+  const [showSettingsModal, setShowSettingsModal] = useState(false)
+  const { formatIndexingTechniqueAndMethod } = useKnowledge()
   const { t } = useTranslation()
 
-  return (
-    <div
-      className={
-        cn(className, s.card,
-          'relative flex cursor-pointer items-center  rounded-xl border border-gray-200 bg-white px-3  py-2.5')
-      }>
-      <div className='flex items-center space-x-2'>
-        <div className={cn(!config.embedding_available && 'opacity-50')}>
-          <TypeIcon type="upload_file" />
-        </div>
-        <div>
-          <div className='mr-1 flex w-[160px] items-center'>
-            <div className={cn('overflow-hidden text-ellipsis whitespace-nowrap text-[13px] font-medium leading-[18px] text-gray-800', !config.embedding_available && 'opacity-50')}>{config.name}</div>
-            {!config.embedding_available && (
-              <Tooltip
-                popupContent={t('dataset.unavailableTip')}
-              >
-                <span className='inline-flex shrink-0 whitespace-nowrap rounded-md border border-gray-200 px-1 text-xs font-normal leading-[18px] text-gray-500'>{t('dataset.unavailable')}</span>
-              </Tooltip>
-            )}
-          </div>
-          <div className={cn('flex max-w-[150px] text-xs text-gray-500', !config.embedding_available && 'opacity-50')}>
-            {formatNumber(config.word_count)} {t('appDebug.feature.dataSet.words')} · {formatNumber(config.document_count)} {t('appDebug.feature.dataSet.textBlocks')}
-          </div>
-        </div>
-      </div>
+  const handleSave = (newDataset: DataSet) => {
+    onSave(newDataset)
+    setShowSettingsModal(false)
+  }
 
-      {!readonly && <RemoveIcon className={`${s.deleteBtn} absolute right-1 top-1/2 translate-y-[-50%]`} onClick={() => onRemove(config.id)} />}
-    </div>
+  const [isDeleting, setIsDeleting] = useState(false)
+
+  const iconInfo = config.icon_info || {
+    icon: '📙',
+    icon_type: 'emoji',
+    icon_background: '#FFF4ED',
+    icon_url: '',
+  }
+
+  return (
+    <div className={cn(
+      'group relative mb-1 flex h-10 w-full cursor-pointer items-center justify-between rounded-lg border-[0.5px] border-components-panel-border-subtle bg-components-panel-on-panel-item-bg px-2 last-of-type:mb-0 hover:bg-components-panel-on-panel-item-bg-hover',
+      isDeleting && 'border-state-destructive-border hover:bg-state-destructive-hover',
+    )}>
+      <div className='flex w-0 grow items-center space-x-1.5'>
+        <AppIcon
+          size='tiny'
+          iconType={iconInfo.icon_type}
+          icon={iconInfo.icon}
+          background={iconInfo.icon_type === 'image' ? undefined : iconInfo.icon_background}
+          imageUrl={iconInfo.icon_type === 'image' ? iconInfo.icon_url : undefined}
+        />
+        <div className='system-sm-medium w-0 grow truncate text-text-secondary' title={config.name}>{config.name}</div>
+      </div>
+      <div className='ml-2 hidden shrink-0 items-center space-x-1 group-hover:flex'>
+        {
+          editable && <ActionButton
+            onClick={(e) => {
+              e.stopPropagation()
+              setShowSettingsModal(true)
+            }}
+          >
+            <RiEditLine className='h-4 w-4 shrink-0 text-text-tertiary' />
+          </ActionButton>
+        }
+        <ActionButton
+          onClick={() => onRemove(config.id)}
+          state={isDeleting ? ActionButtonState.Destructive : ActionButtonState.Default}
+          onMouseEnter={() => setIsDeleting(true)}
+          onMouseLeave={() => setIsDeleting(false)}
+        >
+          <RiDeleteBinLine className={cn('h-4 w-4 shrink-0 text-text-tertiary', isDeleting && 'text-text-destructive')} />
+        </ActionButton>
+      </div>
+      {
+        config.indexing_technique && <Badge
+          className='shrink-0 group-hover:hidden'
+          text={formatIndexingTechniqueAndMethod(config.indexing_technique, config.retrieval_model_dict?.search_method)}
+        />
+      }
+      {
+        config.provider === 'external' && <Badge
+          className='shrink-0 group-hover:hidden'
+          text={t('dataset.externalTag') as string}
+        />
+      }
+      <Drawer isOpen={showSettingsModal} onClose={() => setShowSettingsModal(false)} footer={null} mask={isMobile} panelClassName='mt-16 mx-2 sm:mr-2 mb-3 !p-0 !max-w-[640px] rounded-xl'>
+        <SettingsModal
+          currentDataset={config}
+          onCancel={() => setShowSettingsModal(false)}
+          onSave={handleSave}
+        />
+      </Drawer>
+    </div >
   )
 }
-export default React.memo(CardItem)
+
+export default Item
