+++ web/i18n/fr-FR/app-debug.ts
@@ -105,8 +105,6 @@ const translation = {
     dataSet: {
       title: 'Contexte',
       noData: 'Vous pouvez importer des Connaissances comme contexte',
-      words: 'Mots',
-      textBlocks: 'Blocs de texte',
       selectTitle: 'Sélectionnez la connaissance de référence',
       selected: 'Connaissance sélectionnée',
       noDataSet: 'Aucune connaissance trouvée',
