+++ web/i18n/ru-RU/app-debug.ts
@@ -105,8 +105,6 @@ const translation = {
     dataSet: {
       title: 'Контекст',
       noData: 'Вы можете импортировать знания в качестве контекста',
-      words: 'Слова',
-      textBlocks: 'Текстовые блоки',
       selectTitle: 'Выберите справочные знания',
       selected: 'Знания выбраны',
       noDataSet: 'Знания не найдены',
