+++ web/i18n/sl-SI/dataset.ts
@@ -75,7 +75,6 @@ const translation = {
   intro5: 'se lahko ustvari',
   intro6: ' kot samostojni vtičnik ChatGPT za objavo',
   unavailable: 'Ni na voljo',
-  unavailableTip: 'Vdelani model ni na voljo, potrebno je konfigurirati privzeti vdelani model',
   datasets: 'ZNANJE',
   datasetsApi: 'API DOSTOP',
   externalKnowledgeForm: {
