+++ web/i18n/uk-UA/app-debug.ts
@@ -105,8 +105,6 @@ const translation = {
     dataSet: {
       title: 'Контекст', // Context
       noData: 'Ви можете імпортувати знання як контекст', // You can import Knowledge as context
-      words: 'Слова', // Words
-      textBlocks: 'Текстові блоки', // Text Blocks
       selectTitle: 'Виберіть довідкові знання', // Select reference Knowledge
       selected: 'Знання обрані', // Knowledge selected
       noDataSet: 'Знання не знайдені', // No Knowledge found
