+++ web/i18n/zh-Hant/dataset.ts
@@ -19,7 +19,6 @@ const translation = {
   intro5: '建立',
   intro6: '為獨立的 ChatGPT 外掛釋出使用',
   unavailable: '不可用',
-  unavailableTip: '由於 embedding 模型不可用，需要配置預設 embedding 模型',
   datasets: '知識庫',
   datasetsApi: 'API',
   retrieval: {
